Error bars + housekeeping
Got rid of hover tool on main plot since it's not necessary and was fucking up.

Added a condition in the JS to prevent repeated points on the ln(L) graph.

Added a range to the numpy histogram function for better binning and adjusted x to be the center of the created bins.

Removed some unused variables.

Added error bars to the main plot, will need error for LS method.